Update climate.py
The script is now able to create a seekpoint if the rate of change of either temperature or humidity is too high. The user is able to specify the camera that the seekpoint is created for and also determine what rate of change is too high for temperature and humidity.